requirements.txt
# please install pip requirements using: 
# cat requirements.txt | sed -e '/^\s*#.*$/d' -e '/^\s*$/d' | xargs -n 1 pip install
# otherwise one single compilation error could eat away like 20 minutes for nothing.

# other requirements
yacs
tqdm
rich
sympy
pillow
trimesh
imageio
tensorboard
scikit-image
scikit-learn
torch-tb-profiler

# other requirements not available in conda
smplx
pymcubes
opencv-python

# dev requirements
h5py
ninja
lpips
ujson
pandas
# for unwrapping to get StVK properly
kornia
jupyter
autopep8
pyntcloud
matplotlib
ruamel.yaml
commentjson

# external dependency: easymocap-public (this repo is not publicly available yet)
# for easymocap's vposer: human_pose_prior, this looks like my DotDict implementation... just way more complex
dotmap
# for easymocap loading of SMPL (maybe all pickle loading of SMPL?)
chumpy
mediapipe
func_timeout
pycocotools
spconv-cu116
tensorboardX
git+https://github.com/mmatl/pyopengl
git+https://github.com/nghorbani/human_body_prior
git+https://github.com/zju3dv/EasyMocap

# !: prone to change pytorch version, please install these on demand and manually
# functorch
torch-scatter

# https://storage.googleapis.com/open3d-releases-master/python-wheels/open3d-0.16.0-cp310-cp310-manylinux_2_27_x86_64.whl # TODO: fix the quirky install
# http://www.open3d.org/docs/release/getting_started.html (install the development version from here if the previsou link is expired and python is too new)
# python3.10 support for open3d finally here
# if failed to install open3d (even when installing from latest release?), try to skip it using
# pip install $(grep -v '^ *#\|^open3d' requirements.txt | grep .)
# open3d
open3d

# pip install $(grep -v '^ *#\|^.*open3d\|^torch-sparse\|^torch-geometric\|^.*cholespy\|^.*pytorch3d\|^.*pyopengl' requirements.txt | grep .)
